requirements.txt
crewai>=0.50.0
langchain-groq>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
streamlit>=1.30.0
